% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{buildAcrImage}
\alias{buildAcrImage}
\title{Calls the \code{az acr build} command with name and tag parameters}
\usage{
buildAcrImage()
}
\description{
Calls the \code{az acr build} command with name and tag parameters
}
