% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{buildDockerImage}
\alias{buildDockerImage}
\title{Calls the \code{docker build} command with name and tag parameters}
\usage{
buildDockerImage()
}
\description{
Calls the \code{docker build} command with name and tag parameters
}
