% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_engine.R
\name{create_tmp_and_dockerfile}
\alias{create_tmp_and_dockerfile}
\title{Creates a tmp directory and saves dockerfile there}
\usage{
create_tmp_and_dockerfile(dockerfile_content)
}
\arguments{
\item{dockerfile_content}{content of dockerfile}

\item{folder_name}{the folder in which the file is saved}
}
\description{
Creates a tmp directory and saves dockerfile there
}
