% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{createACI}
\alias{createACI}
\title{create ACI}
\usage{
createACI(resource_group, registry_name, container_name)
}
\arguments{
\item{resource_group}{name of resource group}

\item{registry_name}{name of Azure Container Registry}

\item{container_name}{name of Azure Container Instance}
}
\description{
deploys the Azure Container Instance (container groups)
}
