% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{createACR}
\alias{createACR}
\title{create ACR}
\usage{
createACR(resource_group, registry_name)
}
\arguments{
\item{resource_group}{name of resource group}

\item{registry_name}{name of Azure Container Registry}
}
\description{
deploys the Azure Container Registry
}
