% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{createACRTask}
\alias{createACRTask}
\title{create ACR task}
\usage{
createACRTask(resource_group, registry_name, task_name, secrets)
}
\arguments{
\item{resource_group}{name of resource group}

\item{registry_name}{name of Azure Container Registry}

\item{task_name}{name of Azure Container Registry task}

\item{secrets}{list of username and password for ACR}
}
\description{
creates the Azure Container Registry task that builds the docker image from the git repo
}
