% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{createDockerfile}
\alias{createDockerfile}
\title{Creates a dockerfile}
\usage{
createDockerfile(package_name = "package_name",
  script_name = "script_name")
}
\arguments{
\item{package_name}{name of the package}

\item{script_name}{name of your script to run the package}
}
\description{
Creates a dockerfile
}
