% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{createResourceGroup}
\alias{createResourceGroup}
\title{create Resource Group}
\usage{
createResourceGroup(resource_group, location)
}
\arguments{
\item{resource_group}{name of resource group}

\item{location}{the location of the resource group}
}
\description{
creates the resource group
}
