% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_engine.R
\name{delete_tmp_folder}
\alias{delete_tmp_folder}
\title{Deletes a directory and its content}
\usage{
delete_tmp_folder(folder_name = "tmp")
}
\arguments{
\item{folder_name}{the folder to delete}
}
\description{
Deletes a directory and its content
}
