% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_engine.R
\name{dockerfile_template}
\alias{dockerfile_template}
\title{Creates a dockerfile}
\usage{
dockerfile_template(get_user_input = TRUE)
}
\arguments{
\item{get_user_input}{Will the user input all the specifications or should we use default ones}
}
\description{
Creates a dockerfile
}
