% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{fillJsonTemplate}
\alias{fillJsonTemplate}
\title{Fillout json template}
\usage{
fillJsonTemplate(file_name = "json.json", data)
}
\arguments{
\item{file_name}{name of the json file in the template folder}

\item{data}{list with variables to fill in}
}
\value{
a rendered json object
}
\description{
needs a template and a list of variables that get filled into
}
