% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_deployment_config.R
\name{generate_deployment_config}
\alias{generate_deployment_config}
\title{Generate config file}
\usage{
generate_deployment_config(name, provider, type = "container")
}
\arguments{
\item{name}{Service name}

\item{provider}{Cloud service provider to use. Currently supported "azure"}

\item{type}{Specify if you want to deploy as container or as function}
}
\description{
Generates a config file in current working directory to specify which services
should be created and deployed
}
