% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{getACRSecret}
\alias{getACRSecret}
\title{get ACR secret}
\usage{
getACRSecret(resource_group, registry_name)
}
\arguments{
\item{resource_group}{name of resource group}

\item{registry_name}{name of Azure Container Registry}
}
\value{
a list with the ACR secrets
}
\description{
get the Azure Container Registry username and password
}
