% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{getPackageTitle}
\alias{getPackageTitle}
\title{Helpder function to get the Path of the current package}
\usage{
getPackageTitle()
}
\value{
directory path of package
}
\description{
Helpder function to get the Path of the current package
}
