% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{getToken}
\alias{getToken}
\title{list Resource Groups}
\usage{
getToken(app_name = "test",
  client_id = "04b07795-8ddb-461a-bbee-02f9e1bf7b46",
  client_secret = NULL)
}
\arguments{
\item{app_name}{name of app}

\item{client_id}{client id of Azure CLI}

\item{client_secret}{stays NULL,}
}
\description{
list Resource Groups.
This function might be deleted soon.
}
