% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{getUserCred}
\alias{getUserCred}
\title{Get user credentials via dialogue}
\usage{
getUserCred(cred)
}
\arguments{
\item{cred}{The name of the service.}
}
\value{
user input, i.e. credential
}
\description{
Opens a generic user dialogue to ask for credentials
}
