% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{getUserInputDropdown}
\alias{getUserInputDropdown}
\title{Get user input from dropdown list via dialogue}
\usage{
getUserInputDropdown(choice = c("choice A", "choice B", "choice C"),
  title = "resource")
}
\arguments{
\item{choice}{possible user choice as a vector}

\item{title}{title of dropdown choices}
}
\value{
user input, i.e. chosen option
}
\description{
Opens a dropdown user dialogue to ask for user selection
}
