% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{makeGETRestCall}
\alias{makeGETRestCall}
\title{GET REST call}
\usage{
makeGETRestCall(url)
}
\arguments{
\item{url}{the Azure REST endpoint (= URL)}
}
\value{
the success message
}
\description{
empty GET call template
}
