% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{makePOSTRestCall}
\alias{makePOSTRestCall}
\title{POST REST call}
\usage{
makePOSTRestCall(url, body = "")
}
\arguments{
\item{url}{the Azure REST endpoint (= URL)}

\item{body}{the json body for the REST call}
}
\value{
the success message
}
\description{
empty POST call template
}
