% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{makePUTRestCall}
\alias{makePUTRestCall}
\title{PUT REST call}
\usage{
makePUTRestCall(url, body = "")
}
\arguments{
\item{url}{the Azure REST endpoint (= URL)}

\item{body}{the json body for the REST call}
}
\value{
the success message
}
\description{
empty PUT call template
}
