% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{manageACI}
\alias{manageACI}
\title{start or stop ACI}
\usage{
manageACI(resource_group, container_name, action = "stop")
}
\arguments{
\item{resource_group}{name of resource group}

\item{container_name}{name of container group}

\item{action}{start or stop}
}
\description{
REST call for starting and stopping the container group
}
