% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{moveDockerfile}
\alias{moveDockerfile}
\title{Moves the dockerfile one folder level up}
\usage{
moveDockerfile()
}
\description{
Moves the dockerfile one folder level up in order to be have at the package source level
}
