% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{serverless_addin}
\alias{serverless_addin}
\title{This Addin creates an addin for RStudio that creates a dockerfile}
\usage{
serverless_addin()
}
\description{
This will create a dockerfile which includs your R package.
OBSOLETE
}
