% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{triggerACRTask}
\alias{triggerACRTask}
\title{Run ACR task}
\usage{
triggerACRTask(resource_group, registry_name, task_name)
}
\arguments{
\item{resource_group}{name of resource group}

\item{registry_name}{name of Azure Container Registry}

\item{task_name}{name of Azure Container Registry task}
}
\description{
runs the Azure Container Registry task that builds the docker image from the git repo
}
