% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_deployments.R
\name{createACI}
\alias{createACI}
\title{create ACI}
\usage{
createACI(resource_group, registry_name, container_name,
  container_cpu = "1", container_memory = "1", env_data)
}
\arguments{
\item{resource_group}{name of resource group}

\item{registry_name}{name of Azure Container Registry}

\item{container_name}{name of Azure Container Instance}

\item{container_cpu}{CPU cores}

\item{container_memory}{memory in RAM}

\item{env_data}{a list with environment variables}
}
\description{
deploys the Azure Container Instance (container groups)
}
