% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{createDockerfile}
\alias{createDockerfile}
\title{Creates a dockerfile}
\usage{
createDockerfile(package_name = "package_name")
}
\arguments{
\item{package_name}{The name of the package}
}
\description{
Creates a dockerfile
}
