% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_addin.R
\name{getUserInput}
\alias{getUserInput}
\title{Get user input via dialogue}
\usage{
getUserInput(input)
}
\arguments{
\item{input}{user input specifications}
}
\value{
user input, i.e. package name
}
\description{
Opens a generic user dialogue to ask for normal user input
}
